+++ js/data.js
@@ -1,3 +1,5 @@
+import {getRandomArrayElement, getRandomNumber, getRandomCoordinate, shuffle} from './utils';
+
 const TYPES = [
   'palace',
   'flat',
@@ -44,4 +46,36 @@ const LNG_MAX = 139.8;
 
 const SIMILAR_COUNT = 4;
 
-export {TYPES,TITLES, DESCRIPTION, FEATURES, PHOTOS, TIME_CHECKS, LAT_MIN, LAT_MAX, LNG_MIN, LNG_MAX, SIMILAR_COUNT};
+const createAdvertisement = () => {
+  const offerlocation = {
+    lat: getRandomCoordinate(LAT_MIN, LAT_MAX, 5),
+    lng: getRandomCoordinate(LNG_MIN, LNG_MAX, 5),
+  };
+  const featuresRandomIndex = getRandomNumber(1, FEATURES.length - 1);
+  const features = shuffle(FEATURES).slice(0, featuresRandomIndex);
+
+  return {
+    author: {
+      avatar: `img/avatars/user0${getRandomNumber(1, 10)}.png`,
+    },
+    offer: {
+      title: getRandomArrayElement(TITLES),
+      address: `${offerlocation.lat}, ${offerlocation.lng}`,
+      price: getRandomNumber(0, 10000),
+      type: getRandomArrayElement(TYPES),
+      rooms: getRandomNumber(1, 10),
+      guests: getRandomNumber(1, 10),
+      checkin: getRandomArrayElement(TIME_CHECKS),
+      checkout: getRandomArrayElement(TIME_CHECKS),
+      features,
+      description: getRandomArrayElement(DESCRIPTION),
+      photos: getRandomArrayElement(PHOTOS),
+      location: offerlocation,
+    },
+  };
+};
+const similarAdvertisement = new Array(SIMILAR_COUNT)
+  .fill(null)
+  .map(() => createAdvertisement());
+
+export {similarAdvertisement};
